Move setup stuff to trait
