+++ data-prep/dataprep-simple/src/test/scala/de/hpi/isg/dataprep/preparators/ChangeDateFormatTest.scala
@@ -1,49 +1,15 @@
 package de.hpi.isg.dataprep.preparators
 
-import de.hpi.isg.dataprep.DialectBuilder
-import de.hpi.isg.dataprep.components.{Pipeline, Preparation}
-import de.hpi.isg.dataprep.context.DataContext
-import de.hpi.isg.dataprep.load.FlatFileDataLoader
+import de.hpi.isg.dataprep.components.Preparation
 import de.hpi.isg.dataprep.model.repository.ErrorRepository
 import de.hpi.isg.dataprep.model.target.errorlog.ErrorLog
-import de.hpi.isg.dataprep.model.target.system.AbstractPipeline
 import de.hpi.isg.dataprep.preparators.define.AddProperty
 import de.hpi.isg.dataprep.util.DataType
-import org.apache.spark.sql.{Dataset, Row}
-import org.scalatest._
-import org.apache.log4j.Level
-import org.apache.log4j.Logger
 import org.apache.spark.sql.types.{DataTypes, StructField, StructType, Metadata}
 
 import scala.collection.JavaConverters._
 
-
-class ChangeDateFormatTest extends FlatSpec with Matchers with BeforeAndAfterEach with BeforeAndAfterAll {
-    var dataset: Dataset[Row] = _
-    var pipeline: AbstractPipeline = _
-    var dataContext: DataContext = _
-
-    override def beforeAll: Unit = {
-        Logger.getLogger("org").setLevel(Level.OFF)
-        Logger.getLogger("akka").setLevel(Level.OFF)
-
-        val filePath = getClass.getResource("/pokemon.csv").getPath
-        val dialect = new DialectBuilder()
-            .hasHeader(true)
-            .inferSchema(true)
-            .url(filePath)
-            .buildDialect()
-        val dataLoader = new FlatFileDataLoader(dialect)
-        dataContext = dataLoader.load()
-        //        dataContext.getDataFrame().show()
-        super.beforeAll()
-    }
-
-    override def beforeEach: Unit = {
-        pipeline = new Pipeline(dataContext)
-        super.beforeEach()
-    }
-
+class ChangeDateFormatTest extends PreparatorScalaTest {
     "Test" should "do something" in {
         val preparator = new AddProperty("classic", DataType.PropertyType.INTEGER, 4, 8)
 
@@ -76,6 +42,4 @@ class ChangeDateFormatTest extends FlatSpec with Matchers with BeforeAndAfterEac
         // Second test whether the schema is correctly updated.
         trueSchema shouldEqual updatedSchema
     }
-
-
 }
